# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Informe Centro-Oeste - News Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/news:
    get:
      operationId: getPublicNews
      tags: [public]
      summary: List published news
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, most_viewed]
            default: latest
      responses:
        "200":
          description: Paginated news list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNews"
  /public/news/{slug}:
    get:
      operationId: getPublicNewsArticle
      tags: [public]
      summary: Get single published article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article with related news
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /public/latest-news:
    get:
      operationId: getPublicLatestNews
      tags: [public]
      summary: Get latest news
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Latest news
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsCard"
  /public/most-read:
    get:
      operationId: getPublicMostRead
      tags: [public]
      summary: Get most read articles
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Most read news
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsCard"
  /public/sidebar-news:
    get:
      operationId: getPublicSidebarNews
      tags: [public]
      summary: Get sidebar news by category
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Sidebar news
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsCard"
...
  /public/banners:
    get:
      operationId: getPublicBanners
      tags: [public]
      summary: Get active banners
      parameters:
        - name: position
          in: query
          schema:
            type: string
            enum: [TOP, BELOW_PLAYER, SIDEBAR, BETWEEN_SECTIONS, FOOTER]
      responses:
        "200":
          description: Banners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
  /public/search:
    get:
      operationId: searchPublicNews
      tags: [public]
      summary: Search news
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNews"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, isActive, createdAt]
...
    NewsCard:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
          nullable: true
        featuredImage:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        viewCount:
          type: integer
        isFeatured:
          type: boolean
        category:
          $ref: "#/components/schemas/Category"
        author:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required: [id, name]
      required: [id, title, slug, viewCount, isFeatured, category, author]
    ArticleDetail:
      allOf:
        - $ref: "#/components/schemas/NewsCard"
        - type: object
          properties:
            content:
              type: string
            seoTitle:
              type: string
              nullable: true
            seoDescription:
              type: string
              nullable: true
            seoKeywords:
              type: string
              nullable: true
            tags:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  slug:
                    type: string
                required: [id, name, slug]
            related:
              type: array
              items:
                $ref: "#/components/schemas/NewsCard"
          required: [content]
    PaginatedNews:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NewsCard"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        position:
          type: string
          enum: [TOP, BELOW_PLAYER, SIDEBAR, BETWEEN_SECTIONS, FOOTER]
        imageUrl:
          type: string
        targetUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, position, imageUrl, isActive, createdAt]
...